Create Dessert Clicker App Zip File Link
Modify a starter app called DessertClicker to add logging information that's displayed in the Logcat.
Override lifecycle callback methods and log changes to the activity state.
Run the app and note the logging information that appears as the activity is started, stopped, and resumed.
Implement the onSaveInstanceState() method to retain app data that may be lost if the device configuration changes. Add code to restore that data when the app starts again.