Add files via upload
The completed Forage app allows users to keep track of items, food for example, that they've foraged for in nature. This data is persisted between sessions using Room. You'll use your knowledge of Room and performing read, write, update, and delete operations on a database to implement persistence in the Forage app.